requirements.txt
pushbullet.py
python-decouple
gate-api